refactor: maintain pattern of one utility per file
