+++ subgraphs/venus/src/utilities/index.ts
@@ -1,7 +1,5 @@
 export { getTokenPriceCents } from './getTokenPriceCents';
 export { exponentToBigDecimal } from './exponentToBigDecimal';
 export { getUnderlyingPrice } from './getUnderlyingPrice';
-export {
-  valueOrNotAvailableIntIfReverted,
-  valueOrNotAvailableAddressIfReverted,
-} from './valueOrNotAvailableIfReverted';
+export { valueOrNotAvailableAddressIfReverted } from './valueOrNotAvailableAddressIfReverted';
+export { valueOrNotAvailableIntIfReverted } from './valueOrNotAvailableIntIfReverted';
